refactor(io/read): refine the bloom index file
