+++ src/query/storages/fuse/src/pruning/fuse_bloom_pruner.rs
@@ -117,7 +117,7 @@ impl FuseBloomPrunerCreator {
     pub async fn apply(&self, index_location: &Location, index_length: u64) -> Result<bool> {
         // load the relevant index columns
         let maybe_filter = index_location
-            .read_filter(self.dal.clone(), &self.index_columns, index_length)
+            .read_bloom(self.dal.clone(), &self.index_columns, index_length)
             .await;
 
         match maybe_filter {
